docker/worker/requirements.txt
kafka-python-ng==2.2.3
openai>=1.0.0
firebase-admin>=6.0.0
